olist-etl-pipeline/requirements.txt
pyspark==3.5.0
delta-spark==3.1.0
pandas==2.1.0
pyarrow==14.0.0
boto3==1.34.0
python-dotenv==1.0.0
pytest==7.4.0
apache-airflow==2.8.0
kagglehub==0.3.4
duckdb>=1.0.0
